fix: Refactoring additional android class
