+++ src/Uno.Templates/content/unoapp/MyExtensionsApp.1.Mobile/Android/WebAuthenticationBrokerActivity.Android.cs
@@ -1,0 +1,17 @@
+//-:cnd:noEmit
+using Android.App;
+using Android.Widget;
+using Android.OS;
+using Android.Content.PM;
+using Android.Views;
+
+namespace MyExtensionsApp._1.Droid;
+
+[Activity(NoHistory = true, LaunchMode = LaunchMode.SingleTop, Exported = true)]
+[IntentFilter(
+	new[] { Android.Content.Intent.ActionView },
+	Categories = new[] { Android.Content.Intent.CategoryDefault, Android.Content.Intent.CategoryBrowsable },
+	DataScheme = "myprotocol")]
+public class WebAuthenticationBrokerActivity : Uno.AuthenticationBroker.WebAuthenticationBrokerActivityBase
+{
+}
